@@ -9,14 +9,14 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^5.3 | ^7.0",
+        "php": "^7.1.3",
         "ext-dom": "*",
         "ext-json": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "^4.8 | ^6.5",
-        "satooshi/php-coveralls": ">=0.7.1 <2.0",
-        "squizlabs/php_codesniffer": "^2.3"
+        "php-coveralls/php-coveralls": "^2.1",
+        "squizlabs/php_codesniffer": "^3.4.2"
     },
     "autoload": {
         "psr-4": {
